requirements.txt
playwright
requests
beautifulsoup4
pandas
gspread
google-auth
email-validator
dnspython
aiosmtplib
loguru

